+++ freeway-db/src/main/java/com/jujin/freeway/db/BatchQuery.java
@@ -16,8 +16,6 @@ import java.util.Map;
 public interface BatchQuery {
     BatchQuery rows(Object[]... rows);
 
-    BatchQuery rows(List<Object[]> rows);
-
     BatchQuery named(List<Map<String, Object>> rows);
 
     List<ExecuteResult> execute();